+++ src/lib/types/isRegExp.ts
@@ -1,0 +1,15 @@
+export function isRegExp(value: any): value is RegExp {
+  return (
+    Object.prototype.toString.call(value) === "[object RegExp]" ||
+    value instanceof RegExp ||
+    "[object VRegExp]" === Object.prototype.toString.call(value) ||
+    "[object RegExp]" === Object.prototype.toString.call(value)
+  );
+}
+
+// "string" === typeof value ||
+//     value instanceof String ||
+//     String === value.constructor ||
+//     "[object String]" === Object.prototype.toString.call(value)
+
+export default isRegExp;