+++ src/index.ts
@@ -9,6 +9,7 @@ export { isNumber } from "./lib/types/isNumber";
 export { isBoolean } from "./lib/types/isBoolean";
 export { isExtendable } from "./lib/types/isExtendable";
 export { isRegExp } from "./lib/types/isRegExp";
+export { isDate } from "./lib/types/isDate";
 
 // matchers
 export { isSoftMatch } from "./lib/matchers/isSoftMatch";