+++ src/lib/index.ts
@@ -11,6 +11,7 @@ export { isNumber } from "./types/isNumber";
 export { isBoolean } from "./types/isBoolean";
 export { isExtendable } from "./types/isExtendable";
 export { isRegExp } from "./types/isRegExp";
+export { isDate } from "./types/isDate";
 // matchers
 export { isSoftMatch } from "./matchers/isSoftMatch";
 export { isDeepMatch } from "./matchers/isDeepMatch";